+++ tests/integrations/basic-fail/tests/actual_tests/pattern_too_many_arrow_above.rs
@@ -1,0 +1,6 @@
+fn main() {
+    //~vvvvvv ERROR: mismatched types
+    use_unit(1_u32);
+}
+
+fn use_unit(_: ()) {}