+++ tests/integrations/basic/tests/actual_tests/error_above.rs
@@ -1,0 +1,6 @@
+use basic::add;
+
+fn main() {
+    //~v ERROR: mismatched types
+    add("42", 3);
+}